move calcValues to ODPs
