+++ src/meta/assessment/originalDataPoint/index.ts
@@ -6,4 +6,3 @@ export type { ODPReservedYear } from './odpReservedYear'
 export { ODPs } from './odps'
 export type { ODPValidation, ODPValidationNationalClass, ODPValidationYear } from './odpValidation'
 export type { OriginalDataPoint } from './originalDataPoint'
-export { OriginalDataPointValues } from 'meta/assessment/originalDataPoint/originalDataPointValues'
